Extract authority contract boundary
