+++ docs/specs/README.md
@@ -369,7 +369,7 @@ The initial revision-2 registry must cover the currently known pairs before thei
 
 The registry does not itself grant authority. Every effect still requires a current policy allow, matching runtime grant, exact target/version guards, and the owner specification's reducer. Its canonical fixture manifest includes one allow case and at least wrong-actor, missing-guard, widened-effect, and wrong-grant denials for every required pair; detached, duplicate, or unbound fixtures fail the lifecycle gate. Unknown pairs and entries whose owner dependency is not active fail closed. Draft Wave 1+ specifications therefore bind `SPEC-000@2` only after that revision is current; freezing them against revision 1 would create an avoidable amendment cascade. A later vocabulary addition deliberately creates the corresponding SPEC-000 and dependent-revision impact work instead of silently widening an accepted constitution.
 
-Registry validity is design evidence, not proof that the policy oracle implements it. Before SPEC-000 revision 2 may become `implemented`, the deterministic governance validator must execute every registry fixture against the exact `governance/constitution.yaml` digest and emit `governance/generated/authority-vocabulary-conformance.json`. The receipt binds constitution, registry, fixture-manifest and validator digests; records an allow for every permitted actor; records a denial for every mandatory-guard omission, wrong actor, widened effect and wrong grant; proves unknown pairs deny; and contains no skipped case. The same gate structurally closes the complete policy allow-rule set: every mutating or protected allow rule must resolve to one catalog profile and may not add an actor, omit or replace a mandatory guard, widen an effect, change the grant operation, or hide an alternate permissive condition. The only noncatalog allow rules are separately validated nonmutating `read/*` rules, which the runtime boundary rejects for event append. A retained legacy `publish_draft` rule or any other noncatalog state-changing allow fails even when no finite fixture happens to satisfy its condition. CI recomputes the receipt and rejects a stale, partial, or hand-edited result. Draft, architecture-reviewed, and accepted revisions forbid a prospective receipt; implemented, verified, and operational revisions require it. An amended, superseded, or retired revision may retain its last valid receipt as historical evidence but gains no active authority from it; a never-implemented terminal revision need not create one, and any retained receipt must still validate exactly. The lifecycle stage floor therefore prevents any downstream specification from becoming `implemented` while the registry is merely declarative or the effective oracle still denies or over-allows its required pair.
+Registry validity is design evidence, not proof that the policy oracle implements it. Before SPEC-000 revision 2 may become `implemented`, the deterministic governance validator must execute every registry fixture against the exact `governance/constitution.yaml` digest and emit `governance/generated/authority-vocabulary-conformance.json`. The receipt binds constitution, registry, fixture-manifest, and the complete semantic evaluator bundle: the constitution decision oracle plus the closed authority-contract validator. The bundle manifest has sorted path-and-digest components and its own domain-versioned aggregate digest, so unrelated lifecycle orchestration changes do not churn constitutional evidence and neither semantic component can change unnoticed. Receipt verification also requires the bound component bytes to equal the canonical, non-symlink source files backing the executing evaluator; copied fixture roots pass only when their evaluator bytes are exact. The receipt records an allow for every permitted actor; records a denial for every mandatory-guard omission, wrong actor, widened effect and wrong grant; proves unknown pairs deny; and contains no skipped case. The same gate structurally closes the complete policy allow-rule set: every mutating or protected allow rule must resolve to one catalog profile and may not add an actor, omit or replace a mandatory guard, widen an effect, change the grant operation, or hide an alternate permissive condition. The only noncatalog allow rules are separately validated nonmutating `read/*` rules, which the runtime boundary rejects for event append. A retained legacy `publish_draft` rule or any other noncatalog state-changing allow fails even when no finite fixture happens to satisfy its condition. CI recomputes the receipt and rejects a stale, partial, or hand-edited result. Draft, architecture-reviewed, and accepted revisions forbid a prospective receipt; implemented, verified, and operational revisions require it. An amended, superseded, or retired revision may retain its last valid receipt as historical evidence but gains no active authority from it; a never-implemented terminal revision need not create one, and any retained receipt must still validate exactly. The lifecycle stage floor therefore prevents any downstream specification from becoming `implemented` while the registry is merely declarative or the effective oracle still denies or over-allows its required pair. `validate_authority_contract.py` owns these profile, registry, fixture, closure, and receipt semantics; `validate_spec_lifecycle.py` owns the temporal stage gate and consumes that contract without redefining it.
 
 ## Cross-spec system contracts
